+++ src/parser.c
@@ -75,8 +75,8 @@ void processForget(Forget_t *fo, Journal_t** journal_array) {
 
 }
 
-void destroySchema(Journal_t** journal_array, uint64_t relation_count){
-	uint64_t i;
+void destroySchema(Journal_t** journal_array, int relation_count){
+	int i;
 	for(i = 0; i< relation_count; i++){
 		destroyJournal(journal_array[i]);
 	}
