+++ src/parser.h
@@ -125,7 +125,7 @@ void processForget(Forget_t *, Journal_t**);
 
 Forget_t* forgetParser(char **);
 
-void destroySchema(Journal_t**, uint64_t);
+void destroySchema(Journal_t**, int);
 
 
 #endif
